[#102951968] move PowerStatusWrapper to its own file
