+++ src/reyna/Reyna/Power/IPowerStatusWrapper.cs
@@ -1,0 +1,9 @@
+using System.Windows.Forms;
+
+namespace Reyna.Power
+{
+    public interface IPowerStatusWrapper
+    {
+        PowerLineStatus PowerLineStatus { get; }
+    }
+}
