+++ src/reyna/Reyna/Power/PowerManager.cs
@@ -8,27 +8,14 @@ namespace Reyna.Power
     {
         private readonly IPowerStatusWrapper _powerStatusWrapper;
 
-        public PowerManager(IPowerStatusWrapper powerStatusWrapper = null)
+        public PowerManager(IPowerStatusWrapper powerStatusWrapper)
         {
-            _powerStatusWrapper = powerStatusWrapper ?? new PowerStatusWrapper();
+            _powerStatusWrapper = powerStatusWrapper;
         }
 
         public bool IsPowerLineConnected()
         {
             return _powerStatusWrapper.PowerLineStatus.Equals(PowerLineStatus.Online);
-        }
-    }
-
-    public interface IPowerStatusWrapper
-    {
-        PowerLineStatus PowerLineStatus { get; }
-    }
-
-    public class PowerStatusWrapper : IPowerStatusWrapper
-    {
-        public PowerLineStatus PowerLineStatus
-        {
-            get { return SystemInformation.PowerStatus.PowerLineStatus; }
         }
     }
 }
