+++ src/reyna/Reyna/Power/PowerStatusWrapper.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace Reyna.Power
+{
+    public class PowerStatusWrapper : IPowerStatusWrapper
+    {
+        public PowerLineStatus PowerLineStatus
+        {
+            get { return SystemInformation.PowerStatus.PowerLineStatus; }
+        }
+    }
+}
